Fixed various bugs.
The ground on the LED side of the optoisolator was incorrectly wired
to FTDI ground for both pins; this was rectified.

The board was resized to provide more room for wiring and easier
soldering. I also added another mounting hole.

Various changes made to software, such as the beginnings of a test
routine in the firmware and adding GUI support for such.